Fix invalid data models for chart objects in dynamodb dashboard
The System Errors and Write Throttle Events charts had show chart legend
eneabled but a null value the dimension to show in the legend, which caused
validation errors during imports. This change just removes the chart legend
from the two charts. A better fix is likely to configure this properly, but
we'd need data flowing into signalbuddy master org to do that, which we don't
have...