µAPM 2.0: service and endpoint dashboard updates
Added notes about the setup of the OpenTelemetry collector in addition
to the Smart Agent (which is not yet fully deprecated), and made sure
both the host-level and pod-level info charts had that information.

Cleaned up K8s Pod charts to work with Smart Agent v5.3.x and above that
use kubelet-metrics instead of kubelet-stats.