Update olly signalflow for navs + ECS dashboard bug fix
Adjusted o11ySignalflowTemplate for various navs to
bring the signalflow in line with the signalflow used
for the Entity Nav heatmap.

Modified ElastiCache nav to show Running Nodes on olly
landing page.

Fixed bug in ECS (AWS) dashboard so that it now shows up
on the ECS Infra Navigator (underneath the heatmap).