Azure VMs unified instance dashboard variable change
The new Azure VMs unified instance dashboard had a
dashboard variable for azure_resource_name, but this
variable isn't guaranteed to be collected from otel
hosts, and it also was not being aggregated on in
the heatmap signalflow, leading to issues when opening
a chart from the new Entity Nav. Changed the
dashboard variable to azure_resource_id instead to
resolve that issue.