Fix importQualifiers for My Data Center nav
The importQualifiers for the My Data Center nav did
not include "host.name", just "host"...since the
mapping service isn't used on properties in the
importQualifiers when determining whether to import
content or not, customers that only had hosts that
reported "host.name" were not seeing the My Data
Center nav. This commit addresses that issue.